Consolidate file access machinery in one module
