+++ extra_data/reader.py
@@ -42,7 +42,7 @@ from .read_machinery import (
     find_proposal,
 )
 from .run_files_map import RunFilesMap
-from .filecache import open_files_limiter, file_access_registry
+from .file_access import FileAccess
 
 __all__ = [
     'H5File',
@@ -69,212 +69,6 @@ class by_id(_SliceConstructable):
 
 class by_index(_SliceConstructable):
     pass
-
-
-class FileAccess:
-    """Accesses a single Karabo HDF5 file
-
-    Parameters
-    ----------
-    filename: str
-        A path to an HDF5 file
-    """
-    _file = None
-    _format_version = None
-    metadata_fstat = None
-
-    def __new__(cls, filename, _cache_info=None):
-        # Create only one FileAccess for each path, and store it in a registry
-        filename = osp.abspath(filename)
-        inst = file_access_registry.get(filename, None)
-        if inst is None:
-            inst = file_access_registry[filename] = super().__new__(cls)
-
-        return inst
-
-    def __init__(self, filename, _cache_info=None):
-        self.filename = osp.abspath(filename)
-
-        if _cache_info:
-            self.train_ids = _cache_info['train_ids']
-            self.control_sources = _cache_info['control_sources']
-            self.instrument_sources = _cache_info['instrument_sources']
-        else:
-            tid_data = self.file['INDEX/trainId'][:]
-            self.train_ids = tid_data[tid_data != 0]
-
-            self.control_sources, self.instrument_sources = self._read_data_sources()
-
-            # Store the stat of the file as it was when we read the metadata.
-            # This is used by the run files map.
-            self.metadata_fstat = os.stat(self.file.id.get_vfd_handle())
-
-        # {(file, source, group): (firsts, counts)}
-        self._index_cache = {}
-        # {source: set(keys)}
-        self._keys_cache = {}
-        # {source: set(keys)} - including incomplete sets
-        self._known_keys = defaultdict(set)
-
-    @property
-    def file(self):
-        open_files_limiter.touch(self.filename)
-        if self._file is None:
-            self._file = h5py.File(self.filename, 'r')
-
-        return self._file
-
-    def close(self):
-        """Close* the HDF5 file this refers to.
-
-        The file may not actually be closed if there are still references to
-        objects from it, e.g. while iterating over trains. This is what HDF5
-        calls 'weak' closing.
-        """
-        if self._file:
-            self._file = None
-        open_files_limiter.closed(self.filename)
-
-    @property
-    def format_version(self):
-        if self._format_version is None:
-            version_ds = self.file.get('METADATA/dataFormatVersion')
-            if version_ds is not None:
-                self._format_version = version_ds[0].decode('ascii')
-            else:
-                # The first version of the file format had no version number.
-                # Numbering started at 1.0, so we call the first version 0.5.
-                self._format_version = '0.5'
-
-        return self._format_version
-
-    def _read_data_sources(self):
-        control_sources, instrument_sources = set(), set()
-
-        # The list of data sources moved in file format 1.0
-        if self.format_version == '0.5':
-            data_sources_path = 'METADATA/dataSourceId'
-        else:
-            data_sources_path = 'METADATA/dataSources/dataSourceId'
-
-        for source in self.file[data_sources_path][:]:
-            if not source:
-                continue
-            source = source.decode()
-            category, _, h5_source = source.partition('/')
-            if category == 'INSTRUMENT':
-                device, _, chan_grp = h5_source.partition(':')
-                chan, _, group = chan_grp.partition('/')
-                source = device + ':' + chan
-                instrument_sources.add(source)
-                # TODO: Do something with groups?
-            elif category == 'CONTROL':
-                control_sources.add(h5_source)
-            else:
-                raise ValueError("Unknown data category %r" % category)
-
-        return frozenset(control_sources), frozenset(instrument_sources)
-
-    def __hash__(self):
-        return hash(self.filename)
-
-    def __eq__(self, other):
-        return isinstance(other, FileAccess) and (other.filename == self.filename)
-
-    def __repr__(self):
-        return "{}({})".format(type(self).__name__, repr(self.filename))
-
-    def __getstate__(self):
-        """ Allows pickling `FileAccess` instance. """
-        state = self.__dict__.copy()
-        state['_file'] = None
-        return state
-
-    def __getnewargs__(self):
-        return (self.filename,)
-
-    @property
-    def all_sources(self):
-        return self.control_sources | self.instrument_sources
-
-    def get_index(self, source, group):
-        """Get first index & count for a source and for a specific train ID.
-
-        Indices are cached; this appears to provide some performance benefit.
-        """
-        try:
-            return self._index_cache[(source, group)]
-        except KeyError:
-            ix = self._read_index(source, group)
-            self._index_cache[(source, group)] = ix
-            return ix
-
-    def _read_index(self, source, group):
-        """Get first index & count for a source.
-
-        This is 'real' reading when the requested index is not in the cache.
-        """
-        ntrains = len(self.train_ids)
-        ix_group = self.file['/INDEX/{}/{}'.format(source, group)]
-        firsts = ix_group['first'][:ntrains]
-        if 'count' in ix_group:
-            counts = ix_group['count'][:ntrains]
-        else:
-            status = ix_group['status'][:ntrains]
-            counts = np.uint64((ix_group['last'][:ntrains] - firsts + 1) * status)
-        return firsts, counts
-
-    def get_keys(self, source):
-        """Get keys for a given source name
-
-        Keys are found by walking the HDF5 file, and cached for reuse.
-        """
-        try:
-            return self._keys_cache[source]
-        except KeyError:
-            pass
-
-        if source in self.control_sources:
-            group = '/CONTROL/' + source
-        elif source in self.instrument_sources:
-            group = '/INSTRUMENT/' + source
-        else:
-            raise SourceNameError(source)
-
-        res = set()
-
-        def add_key(key, value):
-            if isinstance(value, h5py.Dataset):
-                res.add(key.replace('/', '.'))
-
-        self.file[group].visititems(add_key)
-        self._keys_cache[source] = res
-        return res
-
-    def has_source_key(self, source, key):
-        """Check if the given source and key exist in this file
-
-        This doesn't scan for all the keys in the source, as .get_keys() does.
-        """
-        try:
-            return key in self._keys_cache[source]
-        except KeyError:
-            pass
-
-        if key in self._known_keys[source]:
-            return True
-
-        if source in self.control_sources:
-            path = '/CONTROL/{}/{}'.format(source, key.replace('.', '/'))
-        elif source in self.instrument_sources:
-            path = '/INSTRUMENT/{}/{}'.format(source, key.replace('.', '/'))
-        else:
-            raise SourceNameError(source)
-
-        if path in self.file:
-            self._known_keys[source].add(key)
-            return True
-        return False
 
 
 class DataCollection:
